# prj/coherence-management/src/main/resources/management-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RESTful Management Interface for Oracle Coherence MBeans",
    "description": "RESTful Management Interface for Oracle Coherence MBeans",
    "version": "12.2.1.4.0"
  },
  "schemes": [
    "http",
    "https"
  ],
...
  "paths": {
...
    "/management/coherence/cluster/services/{serviceName}/persistence/snapshots": {
      "get": {
        "tags": [
          "Persistence"
        ],
        "summary": "View Snapshots",
        "description": "Use this endpoint to view a list of snapshot identifiers that are available to recover from.",
        "operationId": "/management/coherence/cluster/services/{serviceName}/persistence/snapshots GET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/serviceName"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful request that returns a list of snapshots",
            "schema": {
              "type": "object",
              "properties": {
                "snapshots": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/management/coherence/cluster/services/{serviceName}/persistence/archives": {
      "get": {
        "tags": [
          "Persistence"
        ],
        "summary": "View Snapshot Archives",
        "description": "Use this endpoint to view a list of archived snapshots for the service.",
        "operationId": "/management/coherence/cluster/services/{serviceName}/persistence/archives GET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/serviceName"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful request that returns a list of archived snapshots",
            "schema": {
              "type": "object",
              "properties": {
                "snapshots": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "parameters": {
...
    "serviceName": {
      "name": "serviceName",
      "in": "path",
      "description": "Name of the cache service.",
      "required": true,
      "type": "string"
    },
...
  }
}